Fixed cups schema

+++ db/migrate/20200410182020_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration[6.0]
   def change
     create_table :ingredients do |t|
       t.integer :quantity
-      t.integer :cups
+      t.float :cups
       t.float :weight_in_mg
       t.float :weight_in_ml
       t.references :recipe, null: false, foreign_key: true
